Utility.jl

This module holds simple functions that do not yet have a better home or alternatives.

Copyright (c) 2022 Mechanomy LLC, released under the MIT license.

Installation

using Pkg
Pkg.add(url="https://github.com/mechanomy/Utility.jl.git")

API

Utility.Utility — Module

Convenience functions not found elsewhere.

Utility.Angle — Type

Union type for angular units.

Utility.chirpLinear — Function

chirpLinear(start::Unitful.Frequency, stop::Unitful.Frequency, duration::Unitful.Time, amplitude::Number, offset::Number=0, phase=0u"°", ts::Unitful.Time = 1e-3u"s", timeStart::Unitful.Time=0u"s")

Creates a chirp signal from start to stop frequencies over duration time on timestep ts with amplitude about offset, shifted by phase and timeStart

Utility.eqTol — Function

eqTol(A::AbstractMatrix, B::AbstractMatrix) -> Bool
eqTol(A::AbstractMatrix, B::AbstractMatrix, tol) -> Bool

Returns true if every element A and B are equal within tol.

Utility.eqTol — Function

eqTol(a::Number, b::Number) -> Bool
eqTol(a::Number, b::Number, tol) -> Bool

Returns true if a == b within tol. Prefer using Base.isapprox

Utility.eqTol — Function

eqTol(a::Unitful.Quantity, b::Unitful.Quantity, tol=1e-3)::Bool

Returns true if a == b within tol. Prefer using Base.isapprox

Utility.iNext — Method

iNext(i::Integer, n::Integer) -> Integer

Given n elements in an array, return the next i+1 element after i, wrapping around 0 if necessary.

Utility.iNext(1,5) == 2
Utility.iNext(5,5) == 1
Utility.iNext(6,2) == 1
Utility.iNext(-6,5) == 0

Utility.iWrap — Method

iWrap(i::Integer, n::Integer) -> Integer

Given n elements in an array, wrap i around to 1 if it exceeds n.

Utility.iWrap(1,5) == 1
Utility.iWrap(5,5) == 5
Utility.iWrap(6,5) == 1
Utility.iWrap(-6,5) == -1

Utility.stringTable — Function

stringTable(strings::Vector{String}, lengthEach::Integer=3, separator::String="")::String Regularizes the strings vector by truncating each string to lengthEach then concatenating into one string, separated by separator.

Utility.stringTable(["abc","defghi","jk","lmnopq"], 4, "|") == " abc|defg|  jk|lmno"
Utility.stringTable(["abcdef","ghi","jkl","mnopq"], 4, "|") == "abcd| ghi| jkl|mnop"

Utility.within — Method

within(low::Number, value::Number, high::Number) -> Bool

Returns true if low <= value <= high.